+++ examples/example/main.rs
@@ -1,8 +1,8 @@
 extern crate gl;
 extern crate sdl2;
 
+use sdl2::event::Event;
 use sdl2::video::GLProfile;
-use sdl2::{event::Event};
 use std::time::Instant;
 
 use egui::{vec2, Color32, Image, Pos2, Rect};
@@ -12,8 +12,9 @@ use egui_sdl2_gl as egui_backend;
 
 const SCREEN_WIDTH: u32 = 800;
 const SCREEN_HEIGHT: u32 = 600;
-const PIC_WIDTH: i32 = 500;
-const PIC_HEIGHT: i32 = 256;
+const PIC_WIDTH: i32 = 320;
+const PIC_HEIGHT: i32 = 192;
+mod triangle;
 
 fn main() {
     let sdl_context = sdl2::init().unwrap();
@@ -47,7 +48,7 @@ fn main() {
     let mut event_pump = sdl_context.event_pump().unwrap();
     let pixels_per_point = 96f32 / video_subsystem.display_dpi(0).unwrap().0;
     let (width, height) = window.size();
-    
+
     let mut egui_input_state = egui_backend::EguiInputState::new(egui::RawInput {
         screen_rect: Some(Rect::from_min_size(
             Pos2::new(0f32, 0f32),
@@ -76,14 +77,30 @@ fn main() {
 
     //Some variables to help draw a sine wave
     let mut sine_shift = 0f32;
-    
+
     let mut amplitude: f32 = 50f32;
-    let mut test_str: String = "A text box to write text in. Cut, copy, paste commands are available.".to_owned();
+    let mut test_str: String =
+        "A text box to write in. Cut, copy, paste commands are available.".to_owned();
+
+    //We will draw a crisp white triangle using OpenGL.
+    let triangle = triangle::Triangle::new();
+    let mut quit = false;
 
     'running: loop {
         egui_input_state.input.time = Some(start_time.elapsed().as_secs_f64());
 
         egui_ctx.begin_frame(egui_input_state.input.take());
+
+        //An example of how OpenGL can be used to draw custom stuff with egui
+        //overlaying it:
+        //First clear the background to something nice.
+        unsafe {
+            // Clear the screen to black
+            gl::ClearColor(0.3, 0.6, 0.3, 1.0);
+            gl::Clear(gl::COLOR_BUFFER_BIT);
+        }
+        //Then draw our triangle.
+        triangle.draw();
 
         let mut srgba: Vec<Color32> = Vec::new();
         let mut angle = 0f32;
@@ -105,49 +122,51 @@ fn main() {
         //If we weren't updating the texture, this call wouldn't be required.
         painter.update_user_texture_data(chip8_tex_id, &srgba);
 
-        egui::Window::new("Egui with SDL2 events and GL texture").show(&egui_ctx, |ui| {
+        egui::Window::new("Egui with SDL2 and GL").show(&egui_ctx, |ui| {
             //Image just needs a texture id reference, so we just pass it the texture id that was returned to us
             //when we previously initialized the texture.
             ui.add(Image::new(chip8_tex_id, vec2(PIC_WIDTH as f32, PIC_HEIGHT as f32)));
             ui.separator();
-            ui.label("A simple sine wave plotted via some probably dodgy math. The GL texture is dynamically updated and blitted to an egui managed Image.");
+            ui.label("A simple sine wave plotted onto a GL texture then blitted to an egui managed Image.");
             ui.label(" ");
             ui.text_edit_multiline(&mut test_str);
             ui.label(" ");
             ui.add(egui::Slider::f32(&mut amplitude, 0.0..=100.0).text("Amplitude"));
             ui.label(" ");
             if ui.button("Quit").clicked() {
-                std::process::exit(0);
+                quit = true;
             }
         });
 
         let (egui_output, paint_cmds) = egui_ctx.end_frame();
-       
+
         //Handle cut, copy text from egui
         if !egui_output.copied_text.is_empty() {
             egui_backend::copy_to_clipboard(&mut egui_input_state, egui_output.copied_text);
         }
-        
+
         let paint_jobs = egui_ctx.tessellate(paint_cmds);
-        painter.paint_jobs(
-            Color32::from_rgba_premultiplied(0, 0, 255, 0),
-            paint_jobs,
-            &egui_ctx.texture(),
-            pixels_per_point,
-        );
+
+        //Note: passing a bg_color to paint_jobs will clear any previously drawn stuff.
+        //Use this only if egui is being used for all drawing and you aren't mixing your own Open GL
+        //drawing calls with it.
+        //Since we are custom drawing an OpenGL Triangle we don't need egui to clear the background.
+        painter.paint_jobs(None, paint_jobs, &egui_ctx.texture(), pixels_per_point);
+
         window.gl_swap_window();
-        
+
         //Using regular SDL2 event pipeline
         for event in event_pump.poll_iter() {
             match event {
                 Event::Quit { .. } => break 'running,
                 _ => {
-                    
                     egui_backend::input_to_egui(event, &mut egui_input_state);
                 }
             }
         }
         std::thread::sleep(::std::time::Duration::new(0, 1_000_000_000u32 / 60));
+        if quit {
+            break;
+        }
     }
-    painter.cleanup();
 }